+++ team1-api/src/main/java/goorm/deepdive/team1/api/jwt/JwtUtil.java
@@ -1,0 +1,166 @@
+package goorm.deepdive.team1.api.jwt;
+
+import goorm.deepdive.team1.api.jwt.exception.*;
+import goorm.deepdive.team1.api.security.CustomAdminDetails;
+import goorm.deepdive.team1.domain.admin.domain.Admin;
+import goorm.deepdive.team1.domain.admin.domain.Role;
+
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.security.Keys;
+import jakarta.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+
+// 토큰 생성과 검증 로직을 담은 클래스
+@Component
+public class JwtUtil {
+
+    private SecretKey Access_secretKey;
+    private SecretKey refrsh_secretKey;
+
+    @Value("${spring.jwt.secret}")
+    private String secret;
+
+    @Value("${spring.jwt.expiration-seconds}")
+    private long expirationSeconds;
+
+    @Value("${spring.jwt.refresh-secret}")
+    private String refresh_secret;
+
+    @Value("${spring.jwt.expiration-seconds}")
+    private long refresh_expirationSeconds;
+
+
+    @PostConstruct
+    public void init() {
+        this.Access_secretKey = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
+        this.refrsh_secretKey = Keys.hmacShaKeyFor(refresh_secret.getBytes(StandardCharsets.UTF_8));
+    }
+
+    public String createAccessToken(String adminId, String role) {
+        long expirationMs = expirationSeconds * 1000; // 만료시간 (ex expirationSeconds가 600인경우 => 10분)
+        return Jwts.builder()
+                .claim("AdminId", adminId)
+                .claim("role", role)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + expirationMs))
+                .signWith(Access_secretKey, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public String createRefreshToken(String adminId) {
+        long expirationMs = refresh_expirationSeconds * 1000; // 7일
+        return Jwts.builder()
+                .claim("AdminId", adminId)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + expirationMs))
+                .signWith(refrsh_secretKey, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public String getAdminId(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(Access_secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .get("AdminId", String.class);
+    }
+
+    public String getEmail(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(Access_secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .get("email", String.class);
+    }
+
+    public String getRole(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(Access_secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .get("role", String.class);
+    }
+
+    // Refresh AccessToken이 만료되었는지 확인 */
+    public Boolean isExpired(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(Access_secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getExpiration()
+                .before(new Date());
+    }
+
+    // Refresh Token이 만료되었는지 확인 */
+    public Boolean isRefreshTokenExpired(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(refrsh_secretKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getExpiration()
+                .before(new Date());
+    }
+
+    public Authentication getAuthentication(String token) {
+        Claims claims = getClaims(token);
+        String adminId = claims.get("adminId", String.class);
+        String role = claims.get("role", String.class);
+
+        Admin adminEntity = Admin.builder()
+                .id(Long.valueOf(adminId))
+                .role(Role.valueOf(role))
+                .build();
+
+        CustomAdminDetails adminDetails = new CustomAdminDetails(adminEntity);
+        return new UsernamePasswordAuthenticationToken(adminDetails, null, adminDetails.getAuthorities());
+    }
+
+    private Claims getClaims(String token) {
+        // 토큰을 디코딩(파싱)해서 Claims 객체로 변환
+        return Jwts.parserBuilder()
+                .setSigningKey(Access_secretKey)  // Access Token 서명 키 사용
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parserBuilder()
+                    .setSigningKey(secret)
+                    .build()
+                    .parseClaimsJws(token);
+
+            return true; // 유효한 토큰
+        } catch (ExpiredJwtException e) {
+            throw new JwtExpiredException(); // 만료된 토큰 예외 던지기
+        } catch (UnsupportedJwtException e) {
+            throw new JwtUnsupportedException(); // 지원되지 않는 형식
+        } catch (MalformedJwtException e) {
+            throw new JwtMalformedException(); // 잘못된 형식의 토큰
+        } catch (SignatureException e) {
+            throw new JwtSignatureInvalidException(); // 서명 오류
+        } catch (IllegalArgumentException e) {
+            throw new JwtInvalidException(); // 빈 토큰 또는 잘못된 값
+        }
+    }
+
+
+
+
+    public void saveRefreshToken(String email, String refreshToken) {
+        // Redis 또는 DB에 Refresh Token을 저장하는 로직 추가 가능
+    }
+}
